# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIFirstCrypto API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GlobalMarket {
  total_market_cap_usd: number;
  total_volume_usd: number;
  btc_dominance: number;
  active_cryptocurrencies: number;
  market_cap_change_percentage_24h: number;
}

export type CoinSparklineIn7d = {
  price?: number[];
} | null;

export interface Coin {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  total_volume: number;
  price_change_percentage_24h?: number | null;
  sparkline_in_7d?: CoinSparklineIn7d;
}

export interface TrendingCoin {
  id: string;
  name: string;
  symbol: string;
  market_cap_rank?: number | null;
  thumb: string;
  price_btc?: number;
}

export interface TopMovers {
  gainers: Coin[];
  losers: Coin[];
}

export interface MarketSummary {
  summary: string;
  generated_at: string;
}

export interface CoinDetail {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  total_volume: number;
  price_change_percentage_24h?: number | null;
  price_change_percentage_7d?: number | null;
  circulating_supply?: number | null;
  max_supply?: number | null;
  description?: string;
  ai_summary: string;
}

export interface PricePoint {
  timestamp: number;
  price: number;
}

export interface FearGreed {
  value: number;
  value_classification: string;
  timestamp: string;
  updated_at: string;
}

export interface FearGreedHistoryPoint {
  value: number;
  value_classification: string;
  timestamp: string;
}

export type GetCoinsParams = {
  page?: number;
  per_page?: number;
  order?: string;
};

export type GetCoinHistoryParams = {
  days?: number;
};

export type GetFearGreedHistoryParams = {
  limit?: number;
};