# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIFirstCrypto API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coins/{id}:
    get:
      operationId: getCoinDetail
      tags: [coins]
      summary: Coin detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Coin detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinDetail"
  /coins/{id}/history:
    get:
      operationId: getCoinHistory
      tags: [coins]
      summary: Coin price history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Price history data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricePoint"
...
components:
  schemas:
...
    CoinDetail:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        image:
          type: string
        current_price:
          type: number
        market_cap:
          type: number
        market_cap_rank:
          type: integer
        total_volume:
          type: number
        price_change_percentage_24h:
          type: number
          nullable: true
        price_change_percentage_7d:
          type: number
          nullable: true
        circulating_supply:
          type: number
          nullable: true
        max_supply:
          type: number
          nullable: true
        description:
          type: string
        ai_summary:
          type: string
      required:
        - id
        - symbol
        - name
        - image
        - current_price
        - market_cap
        - market_cap_rank
        - total_volume
        - ai_summary
    PricePoint:
      type: object
      properties:
        timestamp:
          type: number
        price:
          type: number
      required:
        - timestamp
        - price
...