# lib/api-zod/src/generated/types/coinDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIFirstCrypto API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoinDetail {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  total_volume: number;
  price_change_percentage_24h?: number | null;
  price_change_percentage_7d?: number | null;
  circulating_supply?: number | null;
  max_supply?: number | null;
  description?: string;
  ai_summary: string;
}